Update current upcaster approach
Update the current upcaster approach to make room for an additional
sample and possible future samples. Do so by moving the current classes
into a json specific package, pointing out these changes are intended
for JacksonSerializers. Also add JavaDoc to describe the intent of the
upcaster and EventUpcasterChainFactory. Further, adjust the
EventUpcasterChainFactory to become a utility class. Lastly, the
FlightDelayedEventUpcaster is renamed to
FlightDelayedEvent0_to_1Upcaster, to clearly reflect what instance is
upcasted

#one-to-many-upcaster