Test module using profiles
By testing this module using the command and rest profile, it became
clear the default XStreamSerializer was causing issues. These only
became apparent as soon as the GiftCardBusinessError contained a
workable toString (for which we converted it to a record). Following
this, the generic serializer has been set to Jackson, solving the
predicaments.

Resolves #233